requirements.txt
alembic==1.6.2
aniso8601==9.0.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.0
marshmallow-sqlalchemy==0.25.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.13
watchdog==2.0.3
Werkzeug==1.0.1
